requirements.txt
pystray>=0.19.5
Pillow>=10.0.0
psutil>=5.9.0
pyinstaller>=6.0.0
